refactor: 폴더 이름 변경 및 파일 이동
- model 폴더 이름 변경
- CarValidator 파일 이동
- TryCountValidator 파일 이동
